gpio: samsung: Convert to platform driver
As the final step of making the driver compatible with multiplatform
builds, this patch converts it to a normal platform driver, which uses
ID table to determine the type of GPIO controller and maps GPIO
registers dynamically based on received resource. This lets us remove
last inclusions of mach/ and plat/ headers.

Signed-off-by: Tomasz Figa <tomasz.figa@gmail.com>