media: s5p-mfc: Move crop register read to seq_done handler
Reading crop registers requires the hardware to be running. Requiring
the hardware to be running during userspace calls is a bit of an
overkill, so just read the values whenever the SEQ_DONE interrupt
signals that they are available and let the G_CROP ioctl just read them
from context data.

While at it, simplify the context state condition a bit.

Signed-off-by: Kuang-che Wu <kcwu@chromium.org>
Signed-off-by: Tomasz Figa <tfiga@chromium.org>