media: s5p-mfc: Set current context in single place
Scattering the assignment over all the code executing hardware commands
is not justified in any way, makes it more difficult to read the code
and easier to make mistakes. In fact there was already one code path
leading to driver loop trying to run the same context again and again
if issuing command fails.

This patch fixes the problem by setting current context as soon as it is
selected for next hardware run.

Signed-off-by: Tomasz Figa <tfiga@chromium.org>