media: s5p-mfc: Consolidate interrupt handling
In current code, handling context or device wake-up,
gating clocks, releasing hardware locks and trying to schedule next
context is very inconsistent. For some interrupts, it is done inside
functions that handle the state machine logic, while for others it is
inside the main interrupt handler itself. There is even a lot of
inconsistency between particular instances of such code, such as
different order of operations, some operations missing and so on.

Move all this generic code to the end of main interrupt handler to
consolidate the code. This should actually fix a bug when one context
triggering an error would prevent other contexts from running, due to
scheduling function not being called.

Signed-off-by: Tomasz Figa <tfiga@chromium.org>