media: s5p-mfc: Consolidate clock control
Currently the calls to s5p_mfc_clock_{enable,disable}() are sprinkled
almost everywhere thorough the driver. However, there is only a limited
number of places where it is really needed, which are:
 - scheduling a context in s5p_mfc_try_run() - enable the clocks before
   calling into hardware to run the context and disable in case of a
   failure,
 - finishing interrupt handling - disable the clocks after all the
   handling is done and no more hardware register access is needed,
 - watchdog worker - to stop the hardware instantly and proceed with
   cleanup and restart,
 - low level init, deinit, sleep, wakeup code, which do not use context
   scheduling, but instead send commands to hardware directly. They
   handle the clocks explicitly as well.

Anything else was either unneeded at all or needed because of a design
that required hardware to be running unnecessarily, which was fixed by
previous patches. Remove all of those.

Signed-off-by: Tomasz Figa <tfiga@chromium.org>