media: s5p-mfc: Sanitize suspend/resume handling
Current code clears the enter_suspend flag in several arbitrary places
in the code, which can lead to further context scheduling racing with
the suspend handler for the hardware lock and taking time longer than
necessary. However, there is no need to clear the flag until resume
handler is called, so clear it only from there.

Moreover, current resume handler does not schedule next context
exection, which may lead to a case when the operation happening before
the suspend will not continue automatically, until userspace performs an
action that schedules next execution (e.g. QBUF). Fix it by calling
s5p_mfc_try_run() after resuming the hardware.

While at it, simplify the waiting code as well, since there is no need
for a loop when wait_event is used and hardware lock can be used as the
condition rather than the integer interrupt flag, which is going to be
removed in next patches.

Signed-off-by: Tomasz Figa <tfiga@chromium.org>