Revert "[media] media: s5p-mfc fix invalid memory access from s5p_mfc_release()"
Commit d695c12c57fe ("[media] media: s5p-mfc fix invalid memory access
from s5p_mfc_release()") tried to solve a problem of driver's remove
callback being called while userspace still holds references to objects
provided by the driver. However, the only case it solves is when
userspace closes related file descriptors. If it tries to do something
else, such as calling IOCTLs, it will still cause a crash, because the
associated video_device would be freed already by
video_device_release(), which is a simple wrapper over kfree().

Another aspect of the problem is that the commit in question bypassed
cleanup steps necessary to cleanly unload the driver, such as freeing
instance buffers.

Revert the commit in question and add a workaround in terms of
setting suppress_bind_attrs in the platfor_driver struct, so that it is
only possible to trigger the remove callback by module unload, which
cannot happen if userspace holds any references to the module, e.g. open
files.

Signed-off-by: Tomasz Figa <tfiga@chromium.org>