media: s5p-mfc: Use dedicated flag for context error state
There are many places in current code where context state is changed
unconditionally. Moreover, such places can be executed even if the
context is in MFCINST_ERROR state, which means that it encountered an
unrecoverable error and needs to be closed. In such case, the context
would end up being processed by hardware again, leading to undefined
behavior.

One way to fix this would be to make sure that all those places don't
change the state if it's MFCINST_ERROR. However, this would overly
complicate the code, so fix it by moving error into a separate context
flag, which can be changed only by error handling code.

Signed-off-by: Tomasz Figa <tfiga@chromium.org>